Reverify unverified transactions if necessary when blockchain actor idle.

+++ neo/Settings.cs
@@ -14,8 +14,8 @@ namespace Neo
         public string[] SeedList { get; private set; }
         public IReadOnlyDictionary<TransactionType, Fixed8> SystemFee { get; private set; }
         public Fixed8 LowPriorityThreshold { get; private set; }
-        public uint MaxTransactionsPerBlock { get; private set; }
-        public uint MaxFreeTransactionsPerBlock { get; private set; }
+        public int MaxTransactionsPerBlock { get; private set; }
+        public int MaxFreeTransactionsPerBlock { get; private set; }
         public uint SecondsPerBlock { get; private set; }
 
         public static Settings Default { get; private set; }
@@ -35,8 +35,8 @@ namespace Neo
             this.SystemFee = section.GetSection("SystemFee").GetChildren().ToDictionary(p => (TransactionType)Enum.Parse(typeof(TransactionType), p.Key, true), p => Fixed8.Parse(p.Value));
             this.SecondsPerBlock = GetValueOrDefault(section.GetSection("SecondsPerBlock"), 15u, p => uint.Parse(p));
             this.LowPriorityThreshold = GetValueOrDefault(section.GetSection("LowPriorityThreshold"), Fixed8.FromDecimal(0.001m), p => Fixed8.Parse(p));
-            this.MaxTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxTransactionsPerBlock"), (uint) 500, uint.Parse);
-            this.MaxFreeTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxFreeTransactionsPerBlock"), (uint) 20, uint.Parse);
+            this.MaxTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxTransactionsPerBlock"), 500, int.Parse);
+            this.MaxFreeTransactionsPerBlock = GetValueOrDefault(section.GetSection("MaxFreeTransactionsPerBlock"), 20, int.Parse);
         }
 
         public T GetValueOrDefault<T>(IConfigurationSection section, T defaultValue, Func<string, T> selector)
